Added PDF files
We added all the LePDF files.  In order to reduce the filesize we exported a sparcer sample of points in x than the one used for the integration: one every 5, plus all the 20 points closer to x=1 (in order to keep a good description of the valence PDF near x=1).